requirements.txt
adrf==0.1.1
asgiref==3.7.2
attrs==23.1.0
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-etc==1.4.0
django-filter==23.2
django-hitcount==1.3.5
django-ninja==0.22.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
pluggy==1.2.0
pydantic==1.10.12
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
